+++ physicsnemo/nn/__init__.py
@@ -121,6 +121,11 @@ from .module.hpx import (
 from .module.kan_layers import KolmogorovArnoldNetwork
 from .module.mlp_layers import Mlp
 from .module.pade import Pade
+from .module.point_transformer_attention import (
+    LocalPointTransformerBlock,
+    LocalTokenCrossAttentionBlock,
+    ResidualMLP,
+)
 from .module.pooling import AttentionPooling, MeanPooling
 from .module.resample_layers import (
     DownSample2D,